model by LGBM

+++ explore.py
@@ -17,8 +17,8 @@ def compute_user_ques_acc(df_train):
     # Rename the 3rd column
     user_ques_accuracy.rename(columns={'answered_correctly': 'user_ques_acc'}, inplace=True)
     
-    # Cast the 3rd column to float16
-    user_ques_accuracy['user_ques_acc'] = user_ques_accuracy['user_ques_acc'].astype("float16")
+    # Cast the 3rd column to float32
+    user_ques_accuracy['user_ques_acc'] = user_ques_accuracy['user_ques_acc'].astype("float32")
 
     return user_ques_accuracy
 
@@ -30,8 +30,8 @@ def compute_user_bundle_acc(df_train):
     # Rename the 3rd column
     user_bundle_accuracy.rename(columns={'answered_correctly': 'user_bundle_acc'}, inplace=True)
     
-    # Cast the 3rd column to float16
-    user_bundle_accuracy['user_bundle_acc'] = user_bundle_accuracy['user_bundle_acc'].astype("float16")
+    # Cast the 3rd column to float32
+    user_bundle_accuracy['user_bundle_acc'] = user_bundle_accuracy['user_bundle_acc'].astype("float32")
 
     return user_bundle_accuracy
 
@@ -43,8 +43,8 @@ def compute_user_tags_acc(df_train):
     # Rename the 3rd column
     user_tags_accuracy.rename(columns={'answered_correctly': 'user_tags_acc'}, inplace=True)
     
-    # Cast the 3rd column to float16
-    user_tags_accuracy['user_tags_acc'] = user_tags_accuracy['user_tags_acc'].astype("float16")
+    # Cast the 3rd column to float32
+    user_tags_accuracy['user_tags_acc'] = user_tags_accuracy['user_tags_acc'].astype("float32")
 
     return user_tags_accuracy
 
@@ -56,8 +56,8 @@ def compute_user_part_acc(df_train):
     # Rename the 3rd column
     user_part_accuracy.rename(columns={'answered_correctly': 'user_part_acc'}, inplace=True)
     
-    # Cast the 3rd column to float16
-    user_part_accuracy['user_part_acc'] = user_part_accuracy['user_part_acc'].astype("float16")
+    # Cast the 3rd column to float32
+    user_part_accuracy['user_part_acc'] = user_part_accuracy['user_part_acc'].astype("float32")
 
     return user_part_accuracy
 
@@ -69,8 +69,8 @@ def compute_user_tagcount_acc(df_train):
     # Rename the 3rd column
     user_tagcount_accuracy.rename(columns={'answered_correctly': 'user_tagcount_acc'}, inplace=True)
     
-    # Cast the 3rd column to float16
-    user_tagcount_accuracy['user_tagcount_acc'] = user_tagcount_accuracy['user_tagcount_acc'].astype("float16")
+    # Cast the 3rd column to float32
+    user_tagcount_accuracy['user_tagcount_acc'] = user_tagcount_accuracy['user_tagcount_acc'].astype("float32")
 
     return user_tagcount_accuracy
 
@@ -82,8 +82,8 @@ def compute_ques_acc(df_train):
     # Rename the column name
     ques_accuracy.rename(columns={'answered_correctly': 'ques_acc'}, inplace=True)
     
-    # Cast the column to float16
-    ques_accuracy['ques_acc'] = ques_accuracy['ques_acc'].astype("float16")
+    # Cast the column to float32
+    ques_accuracy['ques_acc'] = ques_accuracy['ques_acc'].astype("float32")
 
     return ques_accuracy
